+++ kele.el
@@ -84,6 +84,12 @@ pods."
   :type '(alist :key-type symbol :value-type 'integer)
   :group 'kele)
 
+;; TODO (#80): Display in the `kele-get-mode' header what fields were filtered out
+(defcustom kele-filtered-fields '((metadata managedFields)
+                                  (metadata annotations kubectl.kubernetes.io/last-applied-configuration))
+  "Top-level resource fields to never display, e.g. in `kele-get'."
+  :type '(repeat (repeat symbol)))
+
 (define-error 'kele-cache-lookup-error
   "Kele failed to find the requested resource in the cache.")
 (define-error 'kele-request-error "Kele failed in querying the Kubernetes API")
@@ -722,6 +728,8 @@ throws an error."
            :retrieval-time time)
         (error (signal 'kele-request-error (error-message-string err)))))))
 
+;; TODO (#80): Let's define a struct for these "`kele-get-mode' metadata" so that we
+;; don't have to do a new `defvar' every time we want to add a new field
 (defvar kele--resource-context)
 (defvar kele--resource-retrieval-time)
 
@@ -814,6 +822,8 @@ throws an error."
 (cl-defun kele--render-object (object &optional buffer)
   "Render OBJECT in a buffer as YAML.
 
+Filters out fields according to `kele-filtered-fields'.
+
 If BUFFER is provided, renders into it.  Otherwise, a new buffer
 will be created.
 
@@ -838,10 +848,14 @@ context and namespace in its name."
          (buf (or buffer (get-buffer-create buf-name)))
          (obj (if (kele--resource-container-p object)
                   (kele--resource-container-resource object)
-                object)))
+                object))
+         (filtered-obj (-reduce-from (lambda (o keys)
+                                       (apply #'kele--prune o keys))
+                                     obj
+                                     kele-filtered-fields)))
     (with-current-buffer buf
       (erase-buffer)
-      (insert (yaml-encode obj))
+      (insert (yaml-encode filtered-obj))
       (whitespace-cleanup)
       (goto-char (point-min))
 
@@ -858,6 +872,20 @@ context and namespace in its name."
 
       (kele-get-mode 1))
     (select-window (display-buffer buf))))
+
+(defun kele--prune (alist &rest keys)
+  "Delete the sub-tree of ALIST corresponding to KEYS."
+  (let ((prev)
+        (curr alist))
+    (dolist (key keys)
+      (setq prev curr)
+      (setq curr (if-let ((list-p (listp curr))
+                          (res (assq key curr)))
+                     (cdr res)
+                   nil)))
+    (when curr
+      (assq-delete-all (car (last keys)) prev))
+    alist))
 
 (defvar kele--context-keymap nil
   "Keymap for actions on Kubernetes contexts.