+++ tests/unit/test-kele.el
@@ -10,6 +10,16 @@
 
 (require 'kele)
 
+(describe "kele--prune"
+  (describe "when traversal path does not exist"
+    (it "no-ops"
+      (expect (kele--prune '((foo . 1)
+                             (bar . ((baz . 2))))
+                           'foo 'baz)
+              :to-equal
+              '((foo . 1)
+                (bar . ((baz . 2))))))))
+
 (describe "kele--with-progress"
   (it "returns the retval of the last evaluated sexp"
     (expect (kele--with-progress "foobar" (= 1 1)) :to-equal t)))
@@ -338,6 +348,17 @@
       (expect (buffer-string) :to-equal "kind: FakeKind
 metadata:
   name: fake-name")))
+
+  (it "respects `kele-filtered-fields'"
+    (setq fake-obj '((kind . "FakeKind")
+                     (metadata . ((name . "fake-name")
+                                  (foo . bar)))))
+    (let ((kele-filtered-fields '((metadata foo))))
+      (with-temp-buffer
+        (kele--render-object fake-obj (current-buffer))
+        (expect (buffer-string) :to-equal "kind: FakeKind
+metadata:
+  name: fake-name"))))
 
   (describe "buffer titling"
     (describe "when input is `kele--resource-container'"